Change method return type to size_t
Fixes MSVC warning about conversion from size_t

+++ image_transport/include/image_transport/publisher_plugin.h
@@ -77,7 +77,7 @@ public:
    * \brief Returns the number of subscribers that are currently connected to
    * this PublisherPlugin.
    */
-  virtual uint32_t getNumSubscribers() const = 0;
+  virtual size_t getNumSubscribers() const = 0;
 
   /**
    * \brief Returns the communication topic that this PublisherPlugin will publish on.
